improve stats from run reports
breaks up the runs further down into "positive" (where we expect bugs) and
"negatives" (where we don't expect bugs) items and calculate metrics based on those.

Also  drops the "noise" metric (number of suggestions not matched) which is
not super relevant at this point and it's not standarized (so avg is not very good)
in favor of "suggestions" that were sent.