fix(seer): Fix the logging (#2557)
Our logging was a mess with duplicated logs, hard to read json format,
and complicated implementation.

This PR simplifies that by using the standard logger, with a substring
filter that filters out noisy logs. Especially needed now that we show
these logs in Sentry.

May the screenshots below be all the rest of what I have to say.

## Before:
![CleanShot 2025-05-09 at 15 25
19@2x](https://github.com/user-attachments/assets/69364ac7-ee7e-445b-9a07-64089f99c026)

## After:
![CleanShot 2025-05-09 at 15 23
33@2x](https://github.com/user-attachments/assets/1246b905-4eb9-481e-bcb7-66489543dfd1)

---------

Co-authored-by: getsantry[bot] <66042841+getsantry[bot]@users.noreply.github.com>