chore: prompt change to evaluations (relevant-warnings)
Based on [this deep dive](https://www.notion.so/sentry/4th-full-run-1f18b10e4b5d808ea1f3c68bb2661136) for one of the runs
in the dataset it seems like the evaluator LLM was not so focused on the content when matching bugs, but would use
the location as a criteria too

The small changes in the prompt make it focus more on the content when matching, and not so much on location.
I noticed that 2 matched bugs with low content score were no longer matched, but we kept the one with a high
content score (all 3 had pretty good location scores)

And that's all the 3 bugs we had matched :D so, there you go.