feat(autofix): Run repos with GCS caching (#2550)
This PR introduces caching of git repos in GCS.

When initializing repos, it works in the steps below (bolded steps are
new with this PR):
1. Downloads the repo
1.1. **If GCS caching is enabled and the repo archive exists, downloads
it from GCS**
 1.2. Otherwise Clones a repo using `git clone`
2. Syncs it to the expected commit sha 
**3. If GCS caching is enabled**
 **3.1. If the repo has not been cached before**
   **- Copies the repo into a separate directory**
**- "Prunes" the git directory to be a minimal representation of the
repository**
   **- Creates a tar archive of the repo**
   **- Uploads the tar archive of the repo to GCS**
 **3.2. If the repo already has been cached**
**- Uses the db entry's sha and the new sha to compare the commit
timestamps**
**- If the new synced version has a newer commit time, then we upload it
according to the steps in 3.1**

## Limitation
With this setup enabled, we'll be still doing a slow download on the
first use of a repo, and this is a current limitation until we introduce
background codebase caching.

## Artifacts
A cached repo has 2 artifacts, first the tar archive itself on GCS, and
a db entry for it that contains the current sha.

## Note
Note: GCS Caching with this PR is only going to be enabled for the Seer
project in the Sentry org for testing and validation purposes.

---------

Co-authored-by: Rohan Agarwal <47861399+roaga@users.noreply.github.com>
Co-authored-by: getsantry[bot] <66042841+getsantry[bot]@users.noreply.github.com>