feat(codegen): Support multiple repos (#2569)
This PR de-couples the tools from the type of `context` a little bit. As
a result, multi-repo search is supported for `CodegenContext`, which is
useful for bug prediction.

This doesn't change functionality. Would need to supply
`more_readable_repos` to a `CodegenBaseRequest` to cause multi-repo
search.

This PR can be merged in independently of #2568.