feat(bug-prediction-evals): Allow custom commit_sha (#2652)
a custom `commit_sha`—representing a point before the PR branch head—is
useful for evaluating PRs which had a bug in the middle of the review
process. this lets us add more buggy PRs to our eval dataset, w/ the
only caveat that we technically weren't certain that the PR caused a
prod problem. we're just trusting our knowledge about the PR, which
should be fine