fix(autofix): Handle view directory in claude tools (#2656)
If the claude view command was called with a directory path, our check
was not working correctly. This should fix it and return the tree for
the path. Should fix
[SEER-145](https://sentry.sentry.io/issues/6593399718/).

Tacking on: modify `get_file_contents` in `AutofixContext` to not
unnecessarily call GitHub if the file is a locally created file. Should
fix [SEER-EB](https://sentry.sentry.io/issues/6045489887/).