feat(repo-caching): Real backfill job + sync job (#2659)
Introduces a complete backfill job + a repo syncing job.

- The backfill job runs every -:00 and -:30 minutes and loops through a
backfill cursor to backfill for repos assigned to project preferences.
This backfill job will continue to go through repos and loop around,
creating archives for repos without one.
- The syncing job runs every -:15 and -:45 minutes and will fetch repo
archives that haven't been updated in the last 7 days and will keep them
up to date.


These backfill & syncing jobs run on the Autofix Backfill deployment
which is entirely separate from the main Autofix deployment.

Autofix runs will no longer upload archives to GCS, this is to ensure
that we don't make Autofix tasks on the Autofix deployment run for
longer than they need to. Handling updating of the GCS archives for a
repo will be entirely done by the sync job on the separate deployment.