feat(repo_manager): 30 day TTL for cached repos (#2682)
Implements a 30-day TTL for seer repo archives, counting from the last
time they were used. So stagnant repos get deleted.

We add a `last_downloaded_at` row which is used to inform this TTL and
the Sync.
- The sync job now only updates the archive if the repo has been used
within the last 7 days.
- The TTL will remove repos that haven't been updated in 30 days or
hasn't been used in 30 days, whichever is older.

Also removes all the testing with `caplog` cuz it's flaky and updates
the cursorrules.