chore(anomaly-detection): support for source alert id and type (#2695)
Add support for source_id and source_type (in addition to the existing
alert_id). With this change, an alert can be referred to using the
original `alert_id` or the new `source_id` + `source_type`. The store,
delete and detect APIs can now be called with either of these ids or
both. If the APIs are called with both then the system will prioritize
alert_id over the other id.